[AIRFLOW-3468] Move KnownEventType out of models.py
