+++ airflow/models/__init__.py
@@ -4334,16 +4334,6 @@ class Chart(Base):
         return self.label
 
 
-class KnownEventType(Base):
-    __tablename__ = "known_event_type"
-
-    id = Column(Integer, primary_key=True)
-    know_event_type = Column(String(200))
-
-    def __repr__(self):
-        return self.know_event_type
-
-
 class KnownEvent(Base):
     __tablename__ = "known_event"
